Move context annotation to the `core`
