+++ server/src/main/java/io/spine/server/event/model/StateSubscriberMethod.java
@@ -26,11 +26,11 @@ import io.spine.base.Environment;
 import io.spine.base.EventMessage;
 import io.spine.base.FieldPath;
 import io.spine.base.FieldPaths;
+import io.spine.core.BoundedContext;
 import io.spine.core.BoundedContextName;
 import io.spine.core.ByField;
 import io.spine.core.Subscribe;
 import io.spine.logging.Logging;
-import io.spine.server.annotation.BoundedContext;
 import io.spine.server.model.MessageFilter;
 import io.spine.server.model.Model;
 import io.spine.server.model.declare.ParameterSpec;
