+++ server/src/main/java/io/spine/system/server/package-info.java
@@ -37,6 +37,6 @@ package io.spine.system.server;
 
 import com.google.errorprone.annotations.CheckReturnValue;
 import io.spine.annotation.Internal;
-import io.spine.server.annotation.BoundedContext;
+import io.spine.core.BoundedContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;
