+++ server/src/test/java/io/spine/server/bus/given/stock/package-info.java
@@ -29,6 +29,6 @@
 package io.spine.server.bus.given.stock;
 
 import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.server.annotation.BoundedContext;
+import io.spine.core.BoundedContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;
