+++ server/src/test/java/io/spine/server/given/groups/package-info.java
@@ -33,6 +33,6 @@
 package io.spine.server.given.groups;
 
 import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.server.annotation.BoundedContext;
+import io.spine.core.BoundedContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;
