+++ server/src/test/java/io/spine/server/given/organizations/package-info.java
@@ -31,6 +31,6 @@
 package io.spine.server.given.organizations;
 
 import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.server.annotation.BoundedContext;
+import io.spine.core.BoundedContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;
