+++ server/src/test/java/io/spine/server/route/given/sur/package-info.java
@@ -28,6 +28,6 @@
 package io.spine.server.route.given.sur;
 
 import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.server.annotation.BoundedContext;
+import io.spine.core.BoundedContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;
