+++ server/src/test/java/io/spine/system/server/given/client/package-info.java
@@ -24,6 +24,6 @@
 package io.spine.system.server.given.client;
 
 import com.google.errorprone.annotations.CheckReturnValue;
-import io.spine.server.annotation.BoundedContext;
+import io.spine.core.BoundedContext;
 
 import javax.annotation.ParametersAreNonnullByDefault;
